deepshopping-main/requirements.txt
beautifulsoup4==4.12.3
chardet==5.2.0
google-api-python-client==2.137.0
langchain==0.3.20
langchain-community==0.3.19
langchain-core==0.3.45
langchain-fireworks==0.2.7
langchain-google-genai==2.1.0
langchain-openai==0.3.8
langgraph==0.3.11
markdownify==0.13.1
pandas==2.2.0
paramiko==3.4.0
pydantic==2.10.6
pymupdf==1.24.13
pytube==15.0.0
readabilipy==0.2.0
transformers==4.48.2
youtube-transcript-api==0.6.2
fastapi
uvicorn
httpx
python-dotenv
python-multipart 